fix 10

+++ pr03/ex10.py
@@ -12,10 +12,10 @@ def find_divisors(a, b, c, k):
 
 
 try:
-    a = float(input("Введіть число a: "))
-    b = float(input("Введіть число b: "))
-    c = float(input("Введіть число c: "))
-    k = float(input("Введіть число k: "))
+    a = int(input("Введіть число a: "))
+    b = int(input("Введіть число b: "))
+    c = int(input("Введіть число c: "))
+    k = int(input("Введіть число k: "))
 except ValueError as err:
     raise ValueError(err)
